Move SaveManager to namespace NebulaWorld
